pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "optimus"
version = "1.0"
license = { text = "Apache 2.0" }
authors = [
  { name = "Nicolas Boizard", email = "nicolas.bzrd@gmail.com" },
  { name = "Duarte Alves", email = "duartemalves@tecnico.ulisboa.pt" },
  { name = "Hippolyte Gisserot-Boukhlef", email = "hippolte.gisserotboukhlef@gmail.com" }
]
description = "Library to train encoder models, written in Python using PyTorch and compatible with all GPUs."
readme = "README.md"
requires-python = ">=3.11.0"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent"
]
dependencies = [
  "blobfile==3.0.0",
  "einops==0.8.1",
  "fire==0.6.0",
  "ijson==3.3.0",
  "mosaicml_streaming==0.7.6",
  "numpy==2.2.4",
  "pandas==2.2.3",
  "pyarrow==16.0.0",
  "python_dateutil==2.9.0",
  "tensorboard==2.19.0",
  "tiktoken==0.7.0",
  "torch==2.4.1",
  "transformers==4.53.2",
  "zstandard==0.23.0"
]

[project.optional-dependencies]
kernel = [
  "flash_attn==2.7.4",
  "liger_kernel==0.5.5"
]

[project.urls]
Homepage = "https://github.com/Nicolas-BZRD/EuroBERT"
Issues = "https://github.com/Nicolas-BZRD/EuroBERT/issues"

[tool.setuptools]
license-files = ["LICENSE.txt"]
